refactor: restructure components directory
