+++ client/src/components/Inventory/Bikes.jsx
@@ -1,14 +1,14 @@
 import React, { Fragment, useState } from 'react';
 import { Table, Thead, Tbody, Tr, TableCaption, Heading } from '@chakra-ui/react';
-import Loader from 'components/common/Loader';
-import { getParts } from 'utils/api/parts.js';
+import Loader from 'components/common/Loader.jsx';
+import { getBikes } from 'utils/api/bikes.js';
 import { useQuery } from 'react-query';
-import { StyledTableRow, StyledTableHeader, StyledTableCell } from '../../common/Table.jsx';
+import { StyledTableRow, StyledTableHeader, StyledTableCell } from 'components/common/Table.jsx';
 import { TablePagination } from '@material-ui/core';
-import { NoResultImage } from '../../common/Image.jsx';
+import { NoResultImage } from 'components/common/Image.jsx';
 
-const Parts = () => {
-  const { isLoading, isSuccess, data } = useQuery('parts', getParts);
+const Bikes = () => {
+  const { isLoading, isSuccess, data } = useQuery('bikes', getBikes);
   const [page, setPage] = useState(0);
   const [rowsPerPage, setRowsPerPage] = useState(10);
 
@@ -29,42 +29,39 @@ const Parts = () => {
     return (
       <>
         <Heading size="xl" textAlign="center" mt={5}>
-          No Parts
+          No bikes.
         </Heading>
         <NoResultImage />
       </>
     );
   }
-
   return (
     <>
       <Table minWidth="unset" width="100%" variant="striped" colorScheme="light">
-        <TableCaption placement="top">List of parts</TableCaption>
+        <TableCaption placement="top">List of bike products</TableCaption>
         <Thead>
           <Tr>
             <StyledTableHeader>Name</StyledTableHeader>
             <StyledTableHeader>Description</StyledTableHeader>
-            <StyledTableHeader>Quality</StyledTableHeader>
-            <StyledTableHeader>Type</StyledTableHeader>
+            <StyledTableHeader>Weight Amount</StyledTableHeader>
+            <StyledTableHeader>Weight Type</StyledTableHeader>
+            <StyledTableHeader>Selling Price</StyledTableHeader>
+            <StyledTableHeader>Cost Price</StyledTableHeader>
             <StyledTableHeader>Color</StyledTableHeader>
-            <StyledTableHeader>Finish</StyledTableHeader>
-            <StyledTableHeader>Grade</StyledTableHeader>
-            <StyledTableHeader>Detail</StyledTableHeader>
           </Tr>
         </Thead>
         <Tbody>
           {isSuccess &&
-            data.data.slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage).map((part) => (
-              <Fragment key={part.id}>
+            data.data.slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage).map((bike) => (
+              <Fragment key={bike._id}>
                 <StyledTableRow>
-                  <StyledTableCell>{part.name}</StyledTableCell>
-                  <StyledTableCell>{part.description}</StyledTableCell>
-                  <StyledTableCell>{part.quality}</StyledTableCell>
-                  <StyledTableCell>{part.type}</StyledTableCell>
-                  <StyledTableCell>{part.color}</StyledTableCell>
-                  <StyledTableCell>{part.finish}</StyledTableCell>
-                  <StyledTableCell>{part.grade}</StyledTableCell>
-                  <StyledTableCell>{part.detail}</StyledTableCell>
+                  <StyledTableCell>{bike.name}</StyledTableCell>
+                  <StyledTableCell>{bike.description}</StyledTableCell>
+                  <StyledTableCell>{bike.weightAmount}</StyledTableCell>
+                  <StyledTableCell>{bike.weightType}</StyledTableCell>
+                  <StyledTableCell>{bike.sellingPrice}</StyledTableCell>
+                  <StyledTableCell>{bike.costPrice}</StyledTableCell>
+                  <StyledTableCell>{bike.color}</StyledTableCell>
                 </StyledTableRow>
               </Fragment>
             ))}
@@ -83,4 +80,4 @@ const Parts = () => {
   );
 };
 
-export default Parts;
+export default Bikes;
